trekker ut parsing av skjema til egen util
