+++ tavla/app/_components/TileCard/utils.ts
@@ -1,6 +1,34 @@
 import type { TTransportMode } from 'src/types/graphql-schema'
+import type { TileColumnDB } from 'types/db-types/boards'
 import type { LineWithFrontText } from './types'
 
+export type TileFormValues = {
+    columns: TileColumnDB[]
+    count: number | null
+    offset: number | null
+    displayName: string
+    quayLineKeys: string[]
+}
+
+export function parseTileFormData(data: FormData): TileFormValues {
+    const columns = data.getAll('columns') as TileColumnDB[]
+    data.delete('columns')
+    const count = data.get('count') as number | null
+    data.delete('count')
+    const offset = data.get('offset') as number | null
+    data.delete('offset')
+    const displayName = data.get('displayName') as string
+    data.delete('displayName')
+
+    const quayLineKeys: string[] = []
+    for (const value of data.values()) {
+        quayLineKeys.push(value as string)
+    }
+
+    return { columns, count, offset, displayName, quayLineKeys }
+}
+
+// TODO: kan det under fjernse?
 export function sortLineByPublicCode(
     a: LineWithFrontText,
     b: LineWithFrontText,
